Missed declaring/correcting a few datatypes from int to double for Price. Fixed them in ProductEntity and Product

+++ src/main/java/edu/uark/models/api/Product.java
@@ -58,7 +58,7 @@ public class Product {
 	public double getPrice() {
 		return this.price;
 	}
-	public Product setPrice(int price) {
+	public Product setPrice(double price) {
 		this.price = price;
 		return this;
 	}
